back/pyproject.toml
[project]
name = "alar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "loguru>=0.7.3",
    "python-multipart>=0.0.20",
]

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["D"]
ignore = ["D101", "D102", "D103", "D104", "D105", "D107", "D100"]
pydocstyle.convention = "google"

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[dependency-groups]
dev = ["httpx>=0.28.1", "pytest>=9.0.1", "pytest-cov>=7.0.0"]
